Available tutorial complete.

+++ Tutorial 3/Pokedex/app/src/main/java/dam_48286/pokedex/model/PokemonDetails.kt
@@ -4,7 +4,7 @@ data class PokemonDetails(
     var id: Int,
     var pokemon: Pokemon,
     var description: String,
-    var weight: Float,
-    var height: Float,
+    var weight: Double,
+    var height: Double,
     var types: List<PokemonType>
 )
